+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - small alekseon logo
 - cache tags
+- getRecordCollection() method in Form class
+- filter,sort,select on form records collection by field identifier
+- getData from form record by field identifier
 
 ## [102.3.0] - 2023-03-03
 ### Added